When upserting an element, prefer to patch by ID
This change means that if an ID is provided to upsertElement, it will
prefer to patch by id than slug. This means that you can change how
deterministic slugs are generated inside integrations without disrupting
contract data that is already synced to Jellyfish.
This should have no effect on the current system and is 100% backwards
compatible with existing integrations.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>